scala translates to java
