--- linkis-engineconn-plugins/shell/src/main/java/org/apache/linkis/manager/engineplugin/shell/executor/ShellECTaskInfo2.java
@@ -1,54 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.linkis.manager.engineplugin.shell.executor;
-
-public class ShellECTaskInfo2 {
-    private String taskId;
-    private Process process;
-    private YarnAppIdExtractor yarnAppIdExtractor;
-
-    public  ShellECTaskInfo2(String taskId, Process process, YarnAppIdExtractor yarnAppIdExtractor) {
-        this.taskId = taskId;
-        this.process = process;
-        this.yarnAppIdExtractor = yarnAppIdExtractor;
-    }
-
-    public String getTaskId() {
-        return taskId;
-    }
-
-    public void setTaskId(String taskId) {
-        this.taskId = taskId;
-    }
-
-    public Process getProcess() {
-        return process;
-    }
-
-    public void setProcess(Process process) {
-        this.process = process;
-    }
-
-    public YarnAppIdExtractor getYarnAppIdExtractor() {
-        return yarnAppIdExtractor;
-    }
-
-    public void setYarnAppIdExtractor(YarnAppIdExtractor yarnAppIdExtractor) {
-        this.yarnAppIdExtractor = yarnAppIdExtractor;
-    }
-}
