--- linkis-engineconn-plugins/shell/src/main/scala/org/apache/linkis/manager/engineplugin/shell/conf/ShellEngineConnConf.scala
@@ -1,30 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.linkis.manager.engineplugin.shell.conf
-
-import org.apache.linkis.common.conf.CommonVars
-
-object ShellEngineConnConf {
-
-  val SHELL_ENGINECONN_CONCURRENT_LIMIT: Int =
-    CommonVars[Int]("linkis.engineconn.shell.concurrent.limit", 30).getValue
-
-  val LOG_SERVICE_MAX_THREAD_SIZE: Int =
-    CommonVars("linkis.engineconn.shell.log.max.thread.size", 50).getValue
-
-}
