+++ linkis-engineconn-plugins/shell/src/test/java/org/apache/linkis/manager/engineplugin/shell/TestShellEngineConnPlugin.java
@@ -1,0 +1,33 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.linkis.manager.engineplugin.shell;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+public class TestShellEngineConnPlugin {
+
+  @Test
+  public void testShellEngineConnExecutor() {
+    ShellEngineConnPlugin shellEngineConnPlugin = new ShellEngineConnPlugin();
+    Assertions.assertNotNull(shellEngineConnPlugin.getEngineConnFactory());
+    Assertions.assertNotNull(shellEngineConnPlugin.getEngineConnLaunchBuilder());
+    Assertions.assertNotNull(shellEngineConnPlugin.getEngineResourceFactory());
+    Assertions.assertNotNull(shellEngineConnPlugin.getDefaultLabels());
+  }
+}
